allow for file exclusions when downloading files (e.g. don't download the grid file for an ITP where a final file exists).

+++ admin_tools/download_itp_files.py
@@ -15,24 +15,49 @@ on the FTP site.
 
 
 PRODUCTS = {
-    'final': re.compile(r'itp[0-9]*final.zip'),
-    'cormat': re.compile(r'itp[0-9]*cormat.zip'),
-    'grid': re.compile(r'itp[0-9]*grddata.zip'),
-    'raw': re.compile(r'itp[0-9]*rawmat.zip')
+    'final': re.compile(r'itp([0-9]+)final.zip'),
+    'grid': re.compile(r'itp([0-9]+)grddata.zip'),
+    'cormat': re.compile(r'itp([0-9]+)cormat.zip'),
+    'raw': re.compile(r'itp([0-9]+)rawmat.zip'),
+    'rawlocs': re.compile(r'itp[0-9]*rawlocs.dat')
 }
+
+exclude_pattern = {
+    'grid': PRODUCTS['final'],
+    'raw': PRODUCTS['cormat'],
+}
+
+CHECKSUM = False
 
 
 def download_itp_files(directory, product):
+    assert product in PRODUCTS
+    directory = Path(directory) / product
     print('Starting download')
-    product_filter = PRODUCTS[product]
-    (directory / 'zipped').mkdir(parents=True, exist_ok=True)
+
+    if product != 'rawlocs':
+        (directory / 'zipped').mkdir(parents=True, exist_ok=True)
     with FTPHost('ftp.whoi.edu', 'anonymous', 'guest') as ftp_host:
         ftp_host.use_list_a_option = False
         print('Connecting ftp.whoi.edu')
         ftp_host.chdir('/whoinet/itpdata/')
         print('Retrieving file list')
         files = ftp_host.listdir(ftp_host.curdir)
-        final_files = list(filter(product_filter.match, files))
+
+        product_re = PRODUCTS[product]
+        final_n = []
+        if product in exclude_pattern:
+            file_n = \
+                [m[1] for m in [re.search(product_re, f) for f in files] if m]
+            exclude_re = exclude_pattern[product]
+            exclude_n = \
+                [m[1] for m in [re.search(exclude_re, f) for f in files] if m]
+            file_n = set(file_n) - set(exclude_n)
+            final_files = \
+                [m[0] for m in [re.search(product_re, f) for f in files]
+                 if m and m[1] in file_n]
+        else:
+            final_files = list(filter(product_re.match, files))
         print(f'{len(final_files)} ITP files found')
         ftp_host.download('MD5SUMS', directory / 'md5.txt')
         checksums = parse_checksums(directory / 'md5.txt')
@@ -40,10 +65,12 @@ def download_itp_files(directory, product):
 
 
         for file_name in final_files:
-            local_path = directory / 'zipped' / file_name
-
+            if product != 'rawlocs':
+                local_path = directory / 'zipped' / file_name
+            else:
+                local_path = directory / file_name
             state = ftp_host.download_if_newer(file_name, local_path)
-            if calc_md5(local_path) != checksums[file_name]:
+            if CHECKSUM and calc_md5(local_path) != checksums[file_name]:
                 local_path.unlink()
                 print('{} failed the checksum test. The file has been '
                       'deleted. Please try running this script '
@@ -54,10 +81,12 @@ def download_itp_files(directory, product):
                 new_files = True
 
     if new_files:
-        print('New zip files were downloaded')
+        print('New files were downloaded')
     else:
-        print('All zip files are up to date')
+        print('All files are up to date')
 
+def exclude_by_itp_number():
+    pass
 
 def unzip_files(directory):
     zip_directory = directory / 'zipped'
@@ -91,13 +120,13 @@ def calc_md5(path):
 
 
 if __name__ == '__main__':
-    destination_arg = ''
+    destination = Path('')
     product_arg = 'final'
     if len(sys.argv) > 1:
-        destination_arg = sys.argv[1]
+        destination = Path(sys.argv[1])
     if len(sys.argv) > 2:
         product_arg = sys.argv[2]
         assert product_arg in PRODUCTS.keys()
-    directory = Path(destination_arg) / product_arg
-    download_itp_files(directory, product_arg)
-    unzip_files(directory)
+    download_itp_files(destination, product_arg)
+    if product_arg != 'rawlocs':
+        unzip_files(destination)
